chore: move pipe to middleware lib
